api with db works well and also runs inside docker on windows

+++ GokartAPI/src/main/java/com/gokart/GokartAPI/Contoller/ApiControllers.java
@@ -1,22 +1,21 @@
 package com.gokart.GokartAPI.Contoller;
 
+import com.gokart.GokartAPI.Models.Kart;
 import com.gokart.GokartAPI.Models.User;
+import com.gokart.GokartAPI.Repo.KartRepo;
 import com.gokart.GokartAPI.Repo.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.nio.file.Path;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 public class ApiControllers {
     @Autowired
     private UserRepo userRepo;
-    @GetMapping("/")
-    public String getSchedule(){
-        return "welcome";
-    }
+    @Autowired
+    private KartRepo kartRepo;
+    //Schedule
     @GetMapping("/users")
     public List<User> getUsers(){
         return userRepo.findAll();
@@ -28,15 +27,55 @@ public class ApiControllers {
     }
     @PutMapping("/update/{id}")
     public String updateUser(@RequestBody User user, @PathVariable long id){
-       User searchedUser = userRepo.findById(id).get();
-       searchedUser.setFullName(user.getFullName());
-       userRepo.save(searchedUser);
-       return "successful update";
+        if(userRepo.findById(id).isPresent()){
+            User searchedUser = userRepo.findById(id).get();
+            searchedUser.setFullName(user.getFullName());
+            userRepo.save(searchedUser);
+            return "successful update";
+        }
+        return "failed to update";
     }
     @DeleteMapping("/delete/{id}")
     public String deleteUser(@PathVariable long id){
-        User searchedUser = userRepo.findById(id).get();
-        userRepo.delete(searchedUser);
-        return "Deleted...";
+        if(userRepo.findById(id).isPresent()){
+            User searchedUser = userRepo.findById(id).get();
+            userRepo.delete(searchedUser);
+            return "Deleted...";
+        }
+        return "failed to delete";
+    }
+    //Kart endpoints
+    @GetMapping("/karts")
+    public List<Kart> getKarts() {return kartRepo.findAll();}
+    @GetMapping("/karts/{id}")
+    public Kart getKartById(@PathVariable long id){
+        if(kartRepo.findById(id).isPresent()) {
+            return kartRepo.findById(id).get();
+        }
+        return null;
+    }
+    @PutMapping("/updateKart/{id}")
+    public String updateKart(@RequestBody Kart kart, @PathVariable long id){
+        if(kartRepo.findById(id).isPresent()){
+            Kart searchedKart = kartRepo.findById(id).get();
+            searchedKart.setKartNumber(kart.getKartNumber());
+            kartRepo.save(searchedKart);
+            return "successful update";
+        }
+        return "failed to update";
+    }
+    @DeleteMapping("/deleteKart/{id}")
+    public String deleteKart(@PathVariable long id){
+        if(kartRepo.findById(id).isPresent()){
+            Kart searchedKart = kartRepo.findById(id).get();
+            kartRepo.delete(searchedKart);
+            return "Deleted...";
+        }
+        return "failed to delete";
+    }
+    @PostMapping("/saveKart")
+    public String saveKart(@RequestBody Kart kart){
+        kartRepo.save(kart);
+        return "Saved...";
     }
 }
